backend/requirements.txt
asgiref==3.11.1
Django==4.2.16
django-cors-headers==4.3.1
django-environ==0.11.2
django-extensions==3.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
et_xmlfile==2.0.0
inflection==0.5.1
numpy==1.26.4
openpyxl==3.1.2
packaging==26.0
pandas==2.1.4
psycopg2-binary==2.9.9
PyJWT==2.11.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
